dj_rest_auth/tests/requirements.txt
coveralls==1.11.1
django-allauth[socialaccount]~=65.13.0
djangorestframework-simplejwt~=5.5.1
flake8==7.1.1
pyotp~=2.9.0
responses==0.12.1
unittest-xml-reporting==3.2.0
